requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
django-extensions==3.2.3
django-ninja==1.1.0
psycopg2==2.9.9
pydantic==2.6.2
pydantic_core==2.16.3
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
